requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
httpx
python-dotenv
tenacity
streamlit